Decouple DrosophAnnot and DynamicPose
